// apps/api-server/src/common/enums/module.enum.ts
/**
 * ============================================================================
 * MODULE ENUM - MODULES ACADEMIA HUB
 * ============================================================================
 */

export enum Module {
  // Module 1
  ELEVES = 'ELEVES',
  INSCRIPTIONS = 'INSCRIPTIONS',
  DOCUMENTS_SCOLAIRES = 'DOCUMENTS_SCOLAIRES',
  
  // Module 2
  ORGANISATION_PEDAGOGIQUE = 'ORGANISATION_PEDAGOGIQUE',
  MATERIEL_PEDAGOGIQUE = 'MATERIEL_PEDAGOGIQUE',
  EXAMENS = 'EXAMENS',
  BULLETINS = 'BULLETINS',
  
  // Module 3
  FINANCES = 'FINANCES',
  RECOUVREMENT = 'RECOUVREMENT',
  DEPENSES = 'DEPENSES',
  
  // Module 4
  RH = 'RH',
  PAIE = 'PAIE',
  
  // Module 5
  COMMUNICATION = 'COMMUNICATION',
  
  // Module 6
  PARAMETRES = 'PARAMETRES',
  ANNEES_SCOLAIRES = 'ANNEES_SCOLAIRES',
  
  // Module 7
  ORION = 'ORION',
  
  // Module 8
  QHSE = 'QHSE',
}